- Add further development notes - Add v_2_4 support to restore_metadata.py - Make code lint compliant - Add v_2_4 to unit tests
